Extract menu method to a separate file
